requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiohttp-cors==0.7.0
aiosignal==1.3.2
amazon-textract-caller==0.2.4
amazon-textract-response-parser==1.0.3
amazon-textract-textractor==1.8.5
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.8.0
async-timeout==5.0.1
attrs==24.3.0
beautifulsoup4==4.12.3
boto3==1.35.97
boto3-stubs==1.35.97
botocore==1.35.97
botocore-stubs==1.35.97
cachetools==5.5.0
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
colorful==0.5.6
dateparser==1.2.0
defusedxml==0.7.1
diskcache==5.6.3
distlib==0.3.9
distro==1.9.0
editdistance==0.8.1
fasteners==0.19
filelock==3.16.1
frozenlist==1.5.0
fsspec==2024.2.0
google-api-core==2.24.0
google-auth==2.37.0
googleapis-common-protos==1.66.0
grpcio==1.69.0
guidance==0.1.14
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
Jinja2==3.1.5
jiter==0.8.2
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
langchain==0.3.14
langchain-anthropic==0.3.1
langchain-core==0.3.29
langchain-huggingface==0.1.2
langchain-text-splitters==0.3.5
langsmith==0.2.10
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
mdit-py-plugins==0.4.2
mdurl==0.1.2
memray==1.15.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
mypy-boto3-cloudformation==1.35.93
mypy-boto3-dynamodb==1.35.94
mypy-boto3-ec2==1.35.93
mypy-boto3-lambda==1.35.93
mypy-boto3-rds==1.35.95
mypy-boto3-s3==1.35.93
mypy-boto3-sqs==1.35.93
nanoid==2.0.0
networkx==3.4.2
numpy==1.26.4
openai==1.59.6
opencensus==0.11.4
opencensus-context==0.1.3
ordered-set==4.1.0
orjson==3.10.14
overrides==7.7.0
packaging==24.2
pandas==2.2.3
pdf2image==1.17.0
pillow==11.1.0
platformdirs==4.3.6
prometheus_client==0.21.1
propcache==0.2.1
proto-plus==1.25.0
protobuf==5.29.3
py-spy==0.4.0
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.5
pydantic_core==2.27.2
pydot==3.0.4
pyformlang==1.0.10
Pygments==2.19.1
pyparsing==3.2.1
pypdf==4.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
ray==2.40.0
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
rsa==4.9
s3transfer==0.10.4
safetensors==0.5.2
scikit-learn==1.6.1
scipy==1.15.1
sentence-transformers==3.3.1
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
structlog==24.4.0
sycamore-ai==0.1.30
sympy==1.13.1
tabulate==0.9.0
tenacity==8.5.0
textual==1.0.0
threadpoolctl==3.5.0
tiktoken==0.8.0
tokenizers==0.21.0
torch==2.5.1
tqdm==4.67.1
transformers==4.48.0
types-awscrt==0.23.6
types-s3transfer==0.10.4
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uc-micro-py==1.0.3
urllib3==2.3.0
virtualenv==20.28.1
wrapt==1.17.1
XlsxWriter==3.2.0
yarl==1.18.3
